Add BuilderRegistry class to manage all build steps
This class manages all built builders, so that others can find them when
necessary.

In the previous way, dependencies are on object but not the instance of
a builder. This makes it impossible to get any dynamic information. It
becomes a problem when I try to build runtime libs. E.g. the install_dir
may depends on arch, and not decidable statically.

Change-Id: I5ec7f458737ef2376836c1efb964a67b3789d159
Signed-off-by: mydongistiny <jaysonedson@gmail.com>