Hide the symbols in the builtins library
When the NDK switches to the compiler-rt builtins, it would be
convenient if the symbols were hidden even without passing
-Wl,--exclude-libs. Keeping the symbols out of app DSO dynsym tables
ensures that apps use their own copies of the builtins, not those from
the platform libc.so/libm.so libs.

On arm32 and 32-bit x86 only, the platform needs to continue exporting
some builtins symbols from libc.so and libm.so, so generate a special
version of the builtins with the symbols exported.

Also: stop outputting a couple copies of libunwind into self.toolchain
that I think aren't needed, because they're only used by the consumers
of the finished toolchain build, not by any toolchain builder:
 - a copy in the runtimes_ndk_cxx directory
 - libunwind-exported.a

Test: ./build.py --no-build windows --no-lto --debug
Change-Id: I44ec79728db92c089e2f39823c35b6f97d24c1ab
Signed-off-by: Jason Edson <jaysonedson@gmail.com>