Enable FrameHeaderCache for libunwind-exported.a
libunwind-exported.a is a special build of libunwind that is only
exported from libc.so. For this build, we can assume that the dlpi_adds
and dlpi_subs are always present, which are needed to invalidate the
FrameHeaderCache when a DSO is unloaded.

Bionic's link.h still defines these fields even for low __ANDROID_API__
versions, on the theory that the caller of dl_iterate_phdr should check
the `size` parameter at run-time, but LLVM's libunwind doesn't check
it.

Bug: http://b/153025717
Test: build.py
Change-Id: Ifd751a9fe080286e6f99a645edf1367706c24535
Signed-off-by: Jason Edson <jaysonedson@gmail.com>