Support building clang for musl host
Add support for a --musl flag that configures clang to build for a
x86_64-linux-musl host instead of glibc.

Cherry-pick two upstream patches to fix running tests against musl,
and add a third patch to fix DynamicLibraryTest that is obsoleted by
https://reviews.llvm.org/D129127 that removed the test.

Disable the libcxx tests for now, I have CLs that make them pass
but the right configuration knobs don't exist to make the tests work
for both glibc and musl.

Bug: 235405318
Test: build.py --no-build=windows --no-lto --musl
Change-Id: I26f5d3b1f723d59c55472ab81c5b600a300b7a6a
Signed-off-by: Jason Edson <jaysonedson@gmail.com>