Do not explicitly set the profile runtime's path in instrumented build
Bug: http://b/245614328

This was added in the past but doesn't seem necessary anymore in local
testing.  Will validate with go/abtd.

The path was also incorrect after the library's install location changed
recently.  If we no longer set the path explicitly, we can get rid of
the add_lib_links for stage1.

Test: go/abtd
Change-Id: I61cddf8761a339d5c0f56299fc26f09476c9cd7b
Signed-off-by: Jason Edson <jaysonedson@gmail.com>